remove select! block from worker loop
It became overly complicated since tokio cancels the futures for branches
that aren't reached, and for the blocking Redis request that could mean
jobs are lost in-transit over the network.

Instead, simplify the loop to rely on the blocking operation to prevent
the loop from executing too quickly. Also, reduce the blocking duration
to 1 second so workers are able to respond to shutdown requests faster.